+++ docs/changelog.md
@@ -9,6 +9,9 @@ Requires libvips v8.14.2
 * Ensure use of `flip` operation forces random access read (regression in 0.32.0).
   [#3600](https://github.com/lovell/sharp/issues/3600)
 
+* Ensure `linear` operation works with 16-bit input (regression in 0.31.3).
+  [#3605](https://github.com/lovell/sharp/issues/3605)
+
 ### v0.32.0 - 24th March 2023
 
 * Default to using sequential rather than random access read where possible.